modified database structure, added new db objects

+++ artisan.go
@@ -11,14 +11,17 @@ import (
 func main() {
 
 	kernel.Init()
+	if len(os.Args) < 2 {
+		fmt.Println("Usage: \"artisan <command> [arguments]\"")
+		return
+	}
 	arg1 := os.Args[1]
-	fmt.Println(arg1)
 	if arg1 == "make:migration" {
 		arg2 := os.Args[2]
 		kernel.CreateMigration(arg2)
 	} else if arg1 == "migrate:fresh" {
 		kernel.RunMigrations()
-	} else if arg1 == "make:seeder" {
+	} else if arg1 == "db:seed" {
 		arg2 := os.Args[2]
 		arg3 := os.Args[3]
 		count, err := strconv.Atoi(arg3)
@@ -26,6 +29,10 @@ func main() {
 			log.Fatal("Cannot read count of elements")
 		}
 		kernel.CreateSeeder(arg2, count, os.Args[4:]...)
+	} else if arg1 == "db:triggers" {
+		kernel.CreateFunctions()
+	} else if arg1 == "db:indexes" {
+		kernel.CreateIndexes()
 	}
 
 }
